reworked all routine

+++ inc/philosophers.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   philosophers.h                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jpiquet <jocelyn.piquet1998@gmail.com>     +#+  +:+       +#+        */
+/*   By: amerzone <amerzone@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/15 17:17:36 by jpiquet           #+#    #+#             */
-/*   Updated: 2025/09/11 18:05:35 by jpiquet          ###   ########.fr       */
+/*   Updated: 2025/09/16 19:01:41 by amerzone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,7 +44,7 @@ t_args init_args(char **argv, int argc);
 
 // routine
 
-int		get_real_time();
+size_t	get_real_time();
 void	init_time_philo(t_philo *philo);
 int		check_dead(t_philo *philo);
 void	*routine(void *args);
